scraper/pyproject.toml
[tool.poetry]
name = "scraper"
version = "0.1.0"
description = ""
authors = ["hutanmihai <hutanmihai29@gmail.com>"]
readme = "README.md"
packages = [{ include = "src", from = "." }]

[tool.poetry.dependencies]
python = "3.12.2"
beautifulsoup4 = "^4.12.3"
requests = "^2.31.0"
pillow = "^10.2.0"
lxml = "^5.1.0"
pandas = "^2.2.1"


[tool.poetry.group.dev.dependencies]
black = "24.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
